WEKA: Support more flexible handling of boolean options.
These can use values "on", "off", "yes", "no", "t", "f", "y", and "n".
Also, if the value is separated by a space it can be parsed accordingly,
even if NoOptDefValue is set.  This is a somewhat ambiguous syntax, and
a bit unfortunate, but it is believed that this will be more usable in
the long run.